Split UsageFetcher testing helpers
